Contract v3

+++ src/state.rs
@@ -15,7 +15,7 @@ pub struct State {
     pub jackpot_percentage_reward: u8,
     pub token_holder_percentage_fee_reward: u8,
     pub fee_for_drand_worker_in_percentage: u8,
-    pub prize_rank_winner_percentage: Vec<u8>,
+    pub prize_rank_winner_percentage: Vec<u64>,
     pub poll_default_end_height: u64,
     pub price_per_ticket_to_register: Uint128,
     pub terrand_contract_address: CanonicalAddr,
@@ -50,7 +50,7 @@ pub struct WinnerRewardClaims {
 pub struct DrawOwnState {
     pub jackpot_percentage_reward: u8,
     pub token_holder_percentage_fee_reward: u8,
-    pub prize_rank_winner_percentage: Vec<u8>,
+    pub prize_rank_winner_percentage: Vec<u64>,
     pub price_per_ticket_to_register: Uint128,
 }
